todoapp/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
click==7.1.2
dataclasses==0.7
dj-database-url==0.5.0
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
isort==5.5.1
mypy-extensions==0.4.3
pathspec==0.8.0
python-decouple==3.3
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
